// packages/browser-pool/src/events.d.ts
export declare enum BROWSER_POOL_EVENTS {
    BROWSER_LAUNCHED = "browserLaunched",
    BROWSER_RETIRED = "browserRetired",
    BROWSER_CLOSED = "browserClosed",
    PAGE_CREATED = "pageCreated",
    PAGE_CLOSED = "pageClosed"
}
export declare enum BROWSER_CONTROLLER_EVENTS {
    BROWSER_CLOSED = "browserClosed"
}